[Refactor] #4223 hook-quest.cpp/h に定義されていたグローバル関数をItemEntity のオブジェクトメソッドに繰り込んだ
